Fixed byte -> MB math. Some efficiencies added with less looping and goroutines.

+++ model/folder.go
@@ -2,5 +2,5 @@ package model
 
 type Folder struct {
 	Name        string
-	SizeInBytes float64
+	SizeInBytes int64
 }
